+++ src/utils.ts
@@ -15,3 +15,10 @@ export function toLowerCases(
 
   return value.toLowerCase();
 }
+
+export function toArray<T>(value: T[] | T): T[] {
+  if (Array.isArray(value)) {
+    return value;
+  }
+  return [value];
+}